Move frontend/lib/tests/pages to frontend/lib/pages/tests.
