+++ frontend/lib/pages/tests/access-dates.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { getInitialState } from '../../pages/access-dates';
+import { getInitialState } from '../access-dates';
 import Routes from '../../routes';
 import LetterOfComplaintRoutes from '../../letter-of-complaint';
-import { AppTesterPal } from '../app-tester-pal';
+import { AppTesterPal } from '../../tests/app-tester-pal';
 import { AccessDatesMutation_output } from '../../queries/AccessDatesMutation';
