+++ frontend/lib/pages/tests/example-loading-page.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { AppTesterPal } from "../app-tester-pal";
-import ExampleLoadingPage from "../../pages/example-loading-page";
+import { AppTesterPal } from "../../tests/app-tester-pal";
+import ExampleLoadingPage from "../example-loading-page";
 
 test("example loading page works", () => {
   const pal = new AppTesterPal(
