+++ frontend/lib/pages/tests/onboarding-step-3.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import OnboardingStep3 from '../../pages/onboarding-step-3';
-import { AppTesterPal } from '../app-tester-pal';
+import OnboardingStep3 from '../onboarding-step-3';
+import { AppTesterPal } from '../../tests/app-tester-pal';
 import { OnboardingStep3Mutation_output } from '../../queries/OnboardingStep3Mutation';
-import { escapeRegExp } from '../util';
+import { escapeRegExp } from '../../tests/util';
 import Routes from '../../routes';
 import { getLeaseChoiceLabels } from '../../../../common-data/lease-choices';
